+++ packages/fcl/src/utils/web/strategies/utils/extension.js
